requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.1.0
sqlparse==0.5.3
gunicorn==21.2.0
dj-database-url==2.1.0
django-environ==0.12.0
psycopg2-binary==2.9.9
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0